Вынес методы и данные для прав на редактирование табличных знаний в модуль knowledge_proxy.py (они нужны только совместно)
